+++ src/core/tile.cpp
@@ -36,6 +36,6 @@ int
 TileSet::at(size_t x, size_t y) noexcept {
     size_t i = y * width + x;
     assert_(i < width * height);
-    assert_(i < INT32_MAX);
+    assert_(i < static_cast<size_t>(INT32_MAX));
     return firstGid + static_cast<int>(i);
 }
